+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Only display the overlays for nodes that are in the viewer render path and have their properties panel maximized. Can be disabled in Preferences/Viewer. #744
 - Fix rendering from Python (by calling app.render()) when nodes use Python expressions. #560
 - Fix drawing of closed Bezier overlays. #659
+- RotoPaint does not reconnect to the viewer at each stroke by default. Added a button to the viewer interface to re-enable this behavior. #641
 
 
 ## Version 2.4.2